Revert "Update On-prem. / Cloud infra model. and Add Software Migration Model (Initial model)"
